requirements.txt
django-tastypie==0.9.9
pyyaml